# Increase XLSX reading performance