@developedsoftware 2019-11-07
Can this be merged with master?