@adrilo 2019-11-13
While it works great with the test file, there are some files where it performs worse... That's why I'm not sure whether this should be merged as is