@developedsoftware 2019-11-13
I have had to down grade to 2.7. I actually think the creation of millions of cell objects is what is slowing 3.0 down and would explain why the varied performance benchmarks.

Even if I want the values as arrays - to be backward compatible - spout creates the objects in the background anyway and then loops through each one and calls getValue to cast the row as an array. Which defeats the point of returning them as an array.